move util functions to users.util
