+++ corehq/pillows/synclog.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 from __future__ import print_function
 from corehq.apps.receiverwrapper.util import get_version_and_app_from_build_id
 from corehq.apps.users.models import CouchUser, CommCareUser, WebUser
-from corehq.pillows.utils import update_last_sync, update_latest_builds
+from corehq.apps.users.util import update_latest_builds, update_last_sync
 from corehq.util.doc_processor.interface import BaseDocProcessor, DocumentProcessorController
 from corehq.util.doc_processor.couch import CouchDocumentProvider
